+++ client/src/App.jsx
@@ -1,40 +1,44 @@
-import React from 'react';
-import './App.css'
-import './index.css';
-import Header from './components/App/Header/Header';
-import Questonnaire from './components/App/Questonnaire';
-import { ApolloClient, InMemoryCache, ApolloProvider, createHttpLink } from '@apollo/client';
-import { setContext } from '@apollo/client/link/context';
-import { Route, Routes } from 'react-router-dom'
-
+import React from "react";
+import "./App.css";
+import "./index.css";
+import Header from "./components/App/Header/Nav";
+import Questonnaire from "./components/App/Questonnaire";
+import {
+	ApolloClient,
+	InMemoryCache,
+	ApolloProvider,
+	createHttpLink,
+} from "@apollo/client";
+import { setContext } from "@apollo/client/link/context";
+import { Route, Routes } from "react-router-dom";
 
 const httpLink = createHttpLink({
-  uri: 'http://localhost:3001/graphql',
+	uri: "http://localhost:3001/graphql",
 });
 
 const authLink = setContext((_, { headers }) => {
-  return {
-    headers: {
-      ...headers,
-      authorization: `Bearer ${process.env.REACT_APP_KEY}`, 
-    },
-  };
+	return {
+		headers: {
+			...headers,
+			authorization: `Bearer ${process.env.REACT_APP_KEY}`,
+		},
+	};
 });
 
 const client = new ApolloClient({
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
+	link: authLink.concat(httpLink),
+	cache: new InMemoryCache(),
 });
 
 function App() {
-  return (
-    <ApolloProvider client={client}>
-      <Header />
-      <Routes>
-        <Route path="/" element={<Questonnaire />} />
-      </Routes>
-    </ApolloProvider>
-  );  
+	return (
+		<ApolloProvider client={client}>
+			<Header />
+			<Routes>
+				<Route path="/" element={<Questonnaire />} />
+			</Routes>
+		</ApolloProvider>
+	);
 }
 
-export default App;  
+export default App;
